debug version - added vertexPlaneSplit combi

+++ v2.0/sample.h
@@ -8,5 +8,5 @@ struct Sample {
 
 struct SampleInd {
 	int prim = -1;
-	int raynr = -1;
+	uint64_t raynr = -1;
 };
